// my-courses-lab4-beek1800/src/app/models/Course.ts
import { stringify } from "@angular/compiler/src/util";

export class Course{

    courseCode: string = "courseCode";
    subjectCode: string = "subjectCode";
    level: string = "level";
    name: string = "name";
    points: number = 0;
    institutionCode: string = "institutionCode";
    subject: string = "subject";
}